Changed how dot files work
A class derived from the command that created the output is used for
each part of the dot file.  e.g. the output of the netstat -v command
found in tcpip/tcpip.snap is put into an object of class Netstat_v.
These classes can then further parse their pieces.