Intermediate result changed.
The intermediate result from the loop calling SnapParser was a list of
databases (dbs) but that was changed to be a list of hashes that has the
db, the directory, and an array of alerts.  This list is then put inside
another hash that has entries for :snap_list (the list of the hashes
just mentioned) and its own array called :alerts

Keeping the directory of the snap seemed like it might prove useful
eventually.  The purpose of the alert in the first hash is for any
processing module to add an "alert" to which pertain to that particular
snap.  The other level hash with its alert array is for any alerts which
are for the group of snaps.  e.g. if two SEAs on two VIOS are not set up
correctly, this will likely show up as an entry in the alerts array of
the outer most hash.