Create method names on Items
I plan to redo this in a more modular way but now Item creates a method
'foo' when the key 'foo' is assigned.  The name is modified to be a
legal method name and an underscore is prepended if a preexisting
method exists.  e.g. Class goes to class which ends up as :_class