Somewhat major revamp
All the specialized code has been moved off into classes.  For example,
the special code in snapper.rb to print out the seas was moved out into
the Seas class.  Ruby classes can now add themselves to a list to be
called with each snap.  And they can add to a print list which is walked
through for each snap.