Added orig_key and line_number
orig_key[key] returns the original key.  This brings back the ability to
detect conflicts in keys.  I hope to use it in the output so the
original text can be printed.

line_number is intended to record the relative line number of @text that
the item was first created but currently it does nothing useful at all.