Added attributes to context#output
Also added the ability to pass a "modifier" down so that the backup and
control channel can be marked.  Also "fix" a weakness where a context's
done was called only when it was used in a list.  Now it is called at
the end of the run -- either when the list of items has been printed or,
if the item is not in a list, when the item has completed printing.
This need came up because the last thing the item could output could
have the :nocr attribute and so the first line the next item would print
would have been appended to the last line of the previous item.