Fix lsattr snap_processor
Just for fun, I rewrote the lsattr snap processor so it goes through and
determines where the four columns start.  This fixes the problem where
an attribute with an empty value would use the first word of the
description as its value.

Also updated specs to use before(:context) instead of before(:all).