Filters can now be added before class is created
If Filter.add is called for a class that does not exist yet, it saves
the filter.  Item::inherited calls Filter::load_filters to load the
filters that have been saved for it.  I suppose this could be arranged
differently.  Perhaps the add should be on the Item class and keep all
the magic within Item.