Added CEC and group snaps
Not what I had original planned but the output now has the id to system,
then hostname, then the items added to the print list which currnetly
are the interfaces, SEAs, and unused adapters.  The id to system is not
indented, the hostname is nested once, and the items in the list are
nested twice.  The hostname will be repeated if there are multiple snaps
for the same host.  The code is currently in Batch.  It may get adjusted
at some point.