Redo of how CECs, LPARs, and Snaps work
In Batch, we now create a list of CECs and each CEC has a list of LPARs
and each LPAR has a list of snaps.  These are sorted.  The output is
then nested by each of these concepts with the print items nested inside
the Snap.  The CEC, LPAR, and Snap all give info such as the hostname,
id_to_system, and date the snap was taken.