More OO oriented Print
Moved all the print concepts in Item which included a lot of the Filter
handling and put it into a module.  Moved Filter and Context as
contained classes of the Print module.  Now we have Print::Filters and
Print::Context which makes more sense.

Level 0, 1, 2, and 3 output did not change.  RSpec and Cucumber passes.
yardoc does not complain.