Able to defer lsattr processing
The process_snap of lsattr was consuming a large amount of time (over
half the time) if the snap had 6900 devices which is coming for hdisks.
Item was modified to that a proc could be the value and the proc was run
automatically when the value was retrieved.  This allowed lsattr to
first just add a proc to the entry.  Then if the entry is needed, the
relatively expensive code of parsing the lsattr output could be done.

Also discovered various bugs and fixed them along the way.  The lsattr
processing since the snakecase change was not working.