Various enhancements
Added mode of the etherchannel (Bob's suggestion)

Added number of virtual CPUs and SMT mode to the LPAR output.

Added alerts if 802.3ad ether channel Partner System attribute
in the entstat does not match across all of the adapters in
the ether channel.  Similarly, allert for Partner Operational
Key.  I started adding a testcase for this but it was too much
work for too little gain.

Flag devices which are in Defined state.  SEAs are always
reported even if they are defined and children of containers
are always reported even if they are defined.  This helps
users notice that something is amiss with these devices.