Fix lparstat.out issue
The "Type" in lparstat.out can have the value of smt_threads or not.
The attributes of proc0 is now used if present to determine the number
of smt threads and, as a fall back, the value from the Type field of
lparstat.out if present, and "??" is neither is present.