Added "drops" and "dropped" to default output
Along with overruns, underruns, and errors, any output that has
"dropped" or "drops" in it is also sent to the output for the default
level with the exception of the "Receive side statistics".  These
"drops" are somewhat normal due to the switches sending packets that do
not need to be bridged.