Added ha_mode is nil
If the SEA ha_mode != disabled, and there is no control channel, then
the PVID adapter should have "Control" buffers.  That code already
existed.  But in the case that the PdAt entries are not present,
sea.attrs['ha_mode]' comes up nil implying that ha_mode is set to
disabled because that is the default.  Added that case to the code to
prevent us from verifying the control buffers when we should not.