Added options to process_snap and process_batch
I got tired of the benign alerts for multiple PdDv entries for
adapter/pci/riciop and others so I do not emit them unless the level is
above the default of 3.  This forced me to pass @options to the
process_snap and process_batch routines which seemed like a good idea
anyway.