Added Qlogic nmentdd adapters
The adapter name is on the next line after "Device Type:" so the
DEVICE_TYPE_REGEXP pattern had to be changed to a multi-line pattern.
This hit the vlan match which use to match just the empty string but now
it matches the first line which is the hardware address.