Better exit characteristics
Made quitting out of less before all of the output had been
sent out as well as hitting ^C exit more gracefully.  The
negative is that -K is sent to less which means that any ^C
will exit but I think that is what most people will prefer and
it can be altered by the user setting the LESS environment
variable the way they want it.