Teach Snapper to look for pervasive issues
By adding to the list in tools/issues.txt and running
bundle exec tools/create_issues.rb

issues.rb is created.  The code then uses the two global defined in it
to check the snap to see if the snap is vulnerable to any of the
issues.  Also checks emgr.snap to see if an interim fix is applied.