Introduced Ethernet type
I wanted to give a lot more detailed output for the ethernet adapters
but before they were being displayed with a "Devices" filter.  The
detailed print routine will be in the class rather than in the filter
and needed to introduce the new class.

The output at this commit is unchanged.