Verify version and release in marshaled data
The version and release were added to the dump created with the --dump
option and they are compared with the current version and release.  If
there is a mismatch, the marshaled data is not used.

This prevents various errors when new features are added but the old
dump files did not implement them and the code dies as a result.