Check interface for VEAs to for mtu_bypass
For VEAs pre p8, the VEA will not have the large send flag set in
driver flags.  In this case, we check the interface flags to see if
mtu_bypass has been enabled and then assume the port has large send
set on the interface.  We might could instead parse the ifconfig -a
output which would be more accurate.