Better error handing
For non ParseError errors, we got a weird output that did not have any
of the nice information like what file was being parsed at the time of
the failure.  Now for many of the "FileParser" subclasses, this has
been fixed.  There are a couple exceptions but they can be addressed
individually if necessary.