Better APAR reporting
An APAR can be partially applied.  The broken filesets code needs to
trigger when any part of a broken APAR has been applied and not be
satisfied until ALL the parts of the APAR that resolves the issue has
been applied.  These changes make a step in that direction.  The
instfix_out that a perfpmr has needs to be reviewed when a sample
comes in.