Checkpoint.
The html is working.  I moved to the latest JQuery and JQueryUI.  I am
now able to use JQuery's .attr and .addClass rather than the DOM
native interface.  I've added titles to some objects (ethernet
adapters and the polylines fro VEAs to the virtual switches.  I had to
switch to https to get it to work on Ecurep (mixed media between http
and https upset the browser).  The html file is now closed nicely.